# Use a faster priority queue implementation