--- a/depthfirst.py
+++ b/depthfirst.py
@@ -1,40 +1,40 @@
 from collections import deque
 
 def solve(maze):
-	start = maze.start
-	end = maze.end
-	width = maze.width
-	stack = deque([start])
-	shape = (maze.height, maze.width)
-	prev = [None] * (maze.width * maze.height)
-	visited = [False] * (maze.width * maze.height)
-	count = 0
+    start = maze.start
+    end = maze.end
+    width = maze.width
+    stack = deque([start])
+    shape = (maze.height, maze.width)
+    prev = [None] * (maze.width * maze.height)
+    visited = [False] * (maze.width * maze.height)
+    count = 0
 
-	completed = False
-	while stack:
-		count += 1
-		current = stack.pop()
+    completed = False
+    while stack:
+        count += 1
+        current = stack.pop()
 
-		if current == end:
-			completed = True
-			break
+        if current == end:
+            completed = True
+            break
 
-		visited[current.Position[0] * width + current.Position[1]] = True
+        visited[current.Position[0] * width + current.Position[1]] = True
 
-		#import code
-		#code.interact(local=locals())
+        #import code
+        #code.interact(local=locals())
 
-		for n in current.Neighbours:
-			if n != None:
-				npos = n.Position[0] * width + n.Position[1]
-				if visited[npos] == False:
-					stack.append(n)
-					prev[npos] = current
+        for n in current.Neighbours:
+            if n != None:
+                npos = n.Position[0] * width + n.Position[1]
+                if visited[npos] == False:
+                    stack.append(n)
+                    prev[npos] = current
 
-	path = deque()
-	current = end
-	while (current != None):
-		path.appendleft(current)
-		current = prev[current.Position[0] * width + current.Position[1]]
+    path = deque()
+    current = end
+    while (current != None):
+        path.appendleft(current)
+        current = prev[current.Position[0] * width + current.Position[1]]
 
-	return [path, [count, len(path), completed]]
+    return [path, [count, len(path), completed]]