@jmitchell 2017-02-25
While testing more mazes I found a minor bug with the new priority queue. Please wait to merge until I've submitted a fix.