@jmitchell 2017-02-26
I fixed the bug and profiled more extensively on other inputs. The initial gains I saw on `perfect2k.png` apparently don't generalize. There's more context in the latest commit comment. I anticipate more refactoring will lead to performance improvements.

Fibonacci heaps, as I understand it, are theoretically ideal, but not necessarily in practice. Binary heaps, like `heapq`'s implementation, can represent the tree in a contiguous array of memory, which is better for cache locality. In contrast Fibonacci heaps are less consistently structured, so they necessarily have more layers of indirection. I don't know in this case which approach will prove fastest after optimizations, but I'm really curious!