@jmitchell 2017-02-26
The latest change is a nice speed improvement for all priority queues. It includes changing to the `HeapPQ` implementation. More info in commit message.